Move unit test and update existing test for creating A matrix stub
